Firebase get x item n cart

+++ lib/screens/home/components/admin/delete_item.dart
@@ -2,10 +2,9 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:flutterfire_ui/firestore.dart';
 import 'package:fluttertoast/fluttertoast.dart';
-import 'package:youmart_mobitech/model/product_download.dart';
-import 'package:youmart_mobitech/model/product_upload.dart';
 
 import '../../../../constants.dart';
+import '../../../../model/product_download.dart';
 
 final queryProductDownloadModel = FirebaseFirestore.instance
     .collection('product')
@@ -16,8 +15,10 @@ final queryProductDownloadModel = FirebaseFirestore.instance
       toFirestore: (user, _) => user.toJson(),
     );
 
+FirebaseFirestore firebaseFirestore = FirebaseFirestore.instance;
+
 class DeleteItem extends StatefulWidget {
-  DeleteItem({Key? key}) : super(key: key);
+  const DeleteItem({Key? key}) : super(key: key);
 
   @override
   State<DeleteItem> createState() => _DeleteItemState();
@@ -37,7 +38,7 @@ class _DeleteItemState extends State<DeleteItem> {
   Widget buildList(BuildContext context) =>
       FirestoreQueryBuilder<ProductDownloadModel>(
         query: queryProductDownloadModel,
-        pageSize: 2,
+        pageSize: 10,
         builder: (context, snapshot, _) {
           if (snapshot.isFetching) {
             return const Center(child: CircularProgressIndicator());
@@ -68,18 +69,58 @@ class _DeleteItemState extends State<DeleteItem> {
                 final productData = snapshot.docs[index].data();
                 return Card(
                   child: ListTile(
+                    tileColor: colorBase,
                     leading: Image.network(productData.image),
                     trailing: ElevatedButton(
-                      onPressed: () {
-                        FirebaseFirestore firebaseFirestore =
-                            FirebaseFirestore.instance;
-
-                        firebaseFirestore
-                            .collection("product")
-                            .doc(productData.pid)
-                            .delete();
-                        Fluttertoast.showToast(msg: "Product Deleted");
-                      },
+                      onPressed: () => showDialog<String>(
+                        context: context,
+                        barrierDismissible: true,
+                        builder: (BuildContext context) => AlertDialog(
+                          title: const Text(
+                            'Delete Item',
+                            style: TextStyle(
+                              fontSize: 25,
+                              color: colorAccent,
+                              fontWeight: FontWeight.w700,
+                            ),
+                          ),
+                          content: Text(
+                            'Are you sure you want to delete ${productData.name} from database?',
+                            style: const TextStyle(
+                              color: colorPrimaryDark,
+                              fontWeight: FontWeight.w500,
+                            ),
+                          ),
+                          actions: <Widget>[
+                            TextButton(
+                              onPressed: () => Navigator.pop(context, 'Cancel'),
+                              child: const Text(
+                                'Cancel',
+                                style: TextStyle(
+                                  color: colorPrimary,
+                                ),
+                              ),
+                            ),
+                            TextButton(
+                              onPressed: () {
+                                firebaseFirestore
+                                    .collection("product")
+                                    .doc(productData.pid)
+                                    .delete();
+                                Fluttertoast.showToast(msg: "Product Deleted");
+                              },
+                              child: const Text(
+                                'Delete',
+                                style: TextStyle(
+                                  color: colorAccent,
+                                  fontWeight: FontWeight.w700,
+                                ),
+                              ),
+                            ),
+                          ],
+                          backgroundColor: colorBase,
+                        ),
+                      ),
                       style: ElevatedButton.styleFrom(
                         primary: colorAccent,
                         onPrimary: colorBase,
